moved the fips check to util
