+++ pkg/metadata/vmmd/vm.go
@@ -193,9 +193,9 @@ func (md *VM) ClearIPAddresses() {
 // Generate a new SSH keypair for the vm
 func (md *VM) newSSHKeypair() (string, error) {
 	privKeyPath := path.Join(md.ObjectPath(), fmt.Sprintf(constants.VM_SSH_KEY_TEMPLATE, md.GetUID()))
-	// Use ED25519 instead of RSA for performance (it's equally secure, but a lot faster to generate/authenticate)
+	// TODO: In future versions, let the user specify what key algorithm to use through the API types
 	ssh_key_algorithm := "ed25519"
-	if FIPSEnabled() {
+	if util.FIPSEnabled() {
 		// Use rsa on FIPS machines
 		ssh_key_algorithm = "rsa"
 	}
